+++ HISTORY.rst
@@ -3,6 +3,14 @@
 History
 -------
 
+0.1.3 (2014-07-08)
+~~~~~~~~~~~~~~~~~~
+
+* Block as expected on start, even if already started.
+* Raise runtime error if starting the reader without a message handler.
+* Add on_close signal to the reader.
+* Allow upgrading to tls+snappy or tls+deflate.
+
 0.1.2 (2014-07-08)
 ~~~~~~~~~~~~~~~~~~